Move helper functions into one file
